+++ confidant/settings.py
@@ -1,7 +1,18 @@
 import json
 import importlib
-from os import getenv
 from base64 import b64decode
+
+from confidant.utils.misc import (
+    bool_env,
+    str_env,
+    int_env
+)
+
+# Logging
+LOGGING_MODULE = str_env('LOGGING_MODULE', 'logging')
+
+logging = importlib.import_module(LOGGING_MODULE)
+logger = logging.getLogger(__name__)
 
 from confidant.encrypted_settings import EncryptedSettings
 
@@ -10,64 +21,7 @@ class SettingsError(Exception):
     pass
 
 
-def bool_env(var_name, default=False):
-    """
-    Get an environment variable coerced to a boolean value.
-    Example:
-        Bash:
-            $ export SOME_VAL=True
-        settings.py:
-            SOME_VAL = bool_env('SOME_VAL', False)
-    Arguments:
-        var_name: The name of the environment variable.
-        default: The default to use if `var_name` is not specified in the
-                 environment.
-    Returns: `var_name` or `default` coerced to a boolean using the following
-        rules:
-            "False", "false" or "" => False
-            Any other non-empty string => True
-    """
-    test_val = getenv(var_name, default)
-    # Explicitly check for 'False', 'false', and '0' since all non-empty
-    # string are normally coerced to True.
-    if test_val in ('False', 'false', '0'):
-        return False
-    return bool(test_val)
-
-
-def float_env(var_name, default=0.0):
-    """
-    Get an environment variable coerced to a float value.
-    This has the same arguments as bool_env. If a value cannot be coerced to a
-    float, a ValueError will be raised.
-    """
-    return float(getenv(var_name, default))
-
-
-def int_env(var_name, default=0):
-    """
-    Get an environment variable coerced to an integer value.
-    This has the same arguments as bool_env. If a value cannot be coerced to an
-    integer, a ValueError will be raised.
-    """
-    return int(getenv(var_name, default))
-
-
-def str_env(var_name, default=''):
-    """
-    Get an environment variable as a string.
-    This has the same arguments as bool_env.
-    """
-    return getenv(var_name, default)
-
-
 # Basic setup
-
-# Logging
-LOGGING_MODULE = str_env('LOGGING_MODULE', 'logging')
-
-logging = importlib.import_module(LOGGING_MODULE)
-logger = logging.getLogger(__name__)
 
 # Whether or not Confidant is run in debug mode. Never run confidant in debug
 # mode outside of development!
